+++ x/ibcratelimit/ibc_test.go
@@ -1,0 +1,36 @@
+package ibcratelimit
+
+import "testing"
+
+func TestNewEmitErrorAcknowledgement(t *testing.T) {
+	testCases := []struct {
+		name string
+	}{}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+		})
+	}
+}
+
+func TestEmitIBCErrorEvents(t *testing.T) {
+	testCases := []struct {
+		name string
+	}{}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+		})
+	}
+}
+
+func TestIsAckError(t *testing.T) {
+	testCases := []struct {
+		name string
+	}{}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+		})
+	}
+}
