--- x/ibcratelimit/osmosis/osmoutils/ibc.go
@@ -1,91 +1,0 @@
-package osmoutils
-
-import (
-	"encoding/json"
-	"fmt"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
-	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
-	ibcexported "github.com/cosmos/ibc-go/v6/modules/core/exported"
-)
-
-const IbcAcknowledgementErrorType = "ibc-acknowledgement-error"
-
-// NewEmitErrorAcknowledgement creates a new error acknowledgement after having emitted an event with the
-// details of the error.
-func NewEmitErrorAcknowledgement(ctx sdk.Context, err error, errorContexts ...string) channeltypes.Acknowledgement {
-	EmitIBCErrorEvents(ctx, err, errorContexts)
-
-	return channeltypes.NewErrorAcknowledgement(err)
-}
-
-// NewSuccessAckRepresentingAnError creates a new success acknowledgement that represents an error.
-// This is useful for notifying the sender that an error has occurred in a way that does not allow
-// the received tokens to be reverted (which means they shouldn't be released by the sender's ics20 escrow)
-func NewSuccessAckRepresentingAnError(ctx sdk.Context, err error, errorContent []byte, errorContexts ...string) channeltypes.Acknowledgement {
-	EmitIBCErrorEvents(ctx, err, errorContexts)
-
-	return channeltypes.NewResultAcknowledgement(errorContent)
-}
-
-// EmitIBCErrorEvents Emit and Log errors
-func EmitIBCErrorEvents(ctx sdk.Context, err error, errorContexts []string) {
-	logger := ctx.Logger().With("module", IbcAcknowledgementErrorType)
-
-	attributes := make([]sdk.Attribute, len(errorContexts)+1)
-	attributes[0] = sdk.NewAttribute("error", err.Error())
-	for i, s := range errorContexts {
-		attributes[i+1] = sdk.NewAttribute("error-context", s)
-		logger.Error(fmt.Sprintf("error-context: %v", s))
-	}
-
-	ctx.EventManager().EmitEvents(sdk.Events{
-		sdk.NewEvent(
-			IbcAcknowledgementErrorType,
-			attributes...,
-		),
-	})
-}
-
-// MustExtractDenomFromPacketOnRecv takes a packet with a valid ICS20 token data in the Data field and returns the
-// denom as represented in the local chain.
-// If the data cannot be unmarshalled this function will panic
-func MustExtractDenomFromPacketOnRecv(packet ibcexported.PacketI) string {
-	var data transfertypes.FungibleTokenPacketData
-	if err := json.Unmarshal(packet.GetData(), &data); err != nil {
-		panic("unable to unmarshal ICS20 packet data")
-	}
-
-	var denom string
-	if transfertypes.ReceiverChainIsSource(packet.GetSourcePort(), packet.GetSourceChannel(), data.Denom) {
-		// remove prefix added by sender chain
-		voucherPrefix := transfertypes.GetDenomPrefix(packet.GetSourcePort(), packet.GetSourceChannel())
-
-		unprefixedDenom := data.Denom[len(voucherPrefix):]
-
-		// coin denomination used in sending from the escrow address
-		denom = unprefixedDenom
-
-		// The denomination used to send the coins is either the native denom or the hash of the path
-		// if the denomination is not native.
-		denomTrace := transfertypes.ParseDenomTrace(unprefixedDenom)
-		if denomTrace.Path != "" {
-			denom = denomTrace.IBCDenom()
-		}
-	} else {
-		prefixedDenom := transfertypes.GetDenomPrefix(packet.GetDestPort(), packet.GetDestChannel()) + data.Denom
-		denom = transfertypes.ParseDenomTrace(prefixedDenom).IBCDenom()
-	}
-	return denom
-}
-
-// IsAckError checks an IBC acknowledgement to see if it's an error.
-// This is a replacement for ack.Success() which is currently not working on some circumstances
-func IsAckError(acknowledgement []byte) bool {
-	var ackErr channeltypes.Acknowledgement_Error
-	if err := json.Unmarshal(acknowledgement, &ackErr); err == nil && len(ackErr.Error) > 0 {
-		return true
-	}
-	return false
-}
